strip: "September 2026" ran out of its slot and over the arrow
`.daystrip .month` is a fixed 78px, and that width is load-bearing: "May
2026" and "September 2026" are different widths, so a shrink-to-fit slot slid
the `›` and all thirty-one day cells sideways under the pointer that had just
pressed the arrow.

A fixed slot is also a claim that every label fits it, and the long form does
not. "September 2026" measures about 105px at 14.5/500, so it ran out of its
own button and painted across the arrow beside it — a control with a month
name lying over it.

`monthLabel` is `month: 'short'` now. That is the picker's own vocabulary
rather than a new one: `monthNames` was already short, so the grid said "Sep"
on the pill you press while the slot said "September" about the same month.
Measured across the five months the gate pages through, the abbreviated
labels run 56 to 66px inside a 78px slot.

The button's `aria-label` is built from the same string, and that is the
reason not to keep the long form there: WCAG 2.5.3 asks that the accessible
name CONTAIN the visible label, and "Sep 2026" is not a substring of
"September 2026 — choose another". A long accessible name over a short
visible one is a control a voice-control user cannot say the name of.

`overflow: hidden` with an ellipsis is a guard on the cause rather than the
fix. The label is `toLocaleDateString`'s, so it is the machine's language:
the abbreviation holds for English, and the next locale with a longer short
form would be this same defect again. Contained and ellipsed is a bad label;
spilling across a control is a broken one.

## The gate check had to move with it

`strip-holds` guarded against passing vacuously with "the longest name and
the shortest differ in characters, or nothing was tested" — a proxy for
"their natural widths differ". Every English month abbreviates to three
letters, so that guard can never be satisfied again, and a guard that cannot
be satisfied is a check that reports nothing.

It measures the mechanism instead: the slot's own width, identical across
five pagings. Stricter than the proxy was, and verified by removing
`width: 78px` and running it:

    FAIL  strip-holds  the month's slot resizes with its name —
    ["May 2026","Jun 2026","Jul 2026","Aug 2026","Sep 2026"]
    measured [66,61,56,64,62]

Five to ten pixels, which a character count could never have seen — all five
of those names are eight characters. The vacuity guard is now that the month
actually changed, which is the thing that would silently stop this check
paging at all.

The unit test asserts the short form as a COMPARISON rather than as "Aug
2026": the label is the platform's, so a literal would only hold on an
English machine.

`pnpm verify` and `pnpm verify:rendered`, both green.

+++ src/renderer/history/month.ts
@@ -77,9 +77,27 @@ export function stepMonth(
   return { year: moved.getFullYear(), month: moved.getMonth() }
 }
 
-/** "August 2026", in whatever the platform calls it. */
+/**
+ * "Aug 2026", in whatever the platform calls it.
+ *
+ * SHORT, and the reason is the slot it goes in. `.daystrip .month` is a fixed
+ * 78px so that paging a month does not slide the forward arrow and all
+ * thirty-one day cells sideways under the pointer pressing them — and the long
+ * form does not fit it. "September 2026" measured about 105px at 14.5/500, so
+ * it ran out of its own button and painted over the `›` beside it.
+ *
+ * The abbreviation is the picker's own vocabulary: `monthNames` below is
+ * already `month: 'short'`, so the grid says "Sep" on the pill you press and
+ * the slot said "September" about the same month.
+ *
+ * The button's `aria-label` is built from this too, deliberately. WCAG 2.5.3
+ * asks that the accessible name CONTAIN the visible label, and "Sep 2026" is
+ * not a substring of "September 2026" — so a long accessible name over a short
+ * visible one would be a control a voice-control user cannot say the name of.
+ * One string, both places.
+ */
 export function monthLabel(year: number, month: number): string {
-  return new Date(year, month, 1).toLocaleDateString(undefined, { month: 'long', year: 'numeric' })
+  return new Date(year, month, 1).toLocaleDateString(undefined, { month: 'short', year: 'numeric' })
 }
 
 /** "Thursday, 14 August" — the heading over one day's conversations. */
